// src/app/dashboard/interfaces/lider-response.interface.ts
export interface LiderResponse {
    id:            number;
    nombrepadre:   string;
    idpadre:       number;
    nombre:        string;
    telefono:      string;
    inefrente:     string;
    ocr:           string;
    fecha:         string;
    domilio:       string;
    referenciadom: string;
    colonia:       string;
    seccion:       string;
    programas:     string;
    observaciones: string;
    traslado:      boolean;
}