// src/app/dashboard/interfaces/lider-update.interface.ts
export interface LiderUpdate {
    id_coordinador: number;
    nombre:         string;
    telefono:       string;
    domicilio:      string;
    referenciadom:  string;
    cve_colonia:    number;
    seccion:        string;
    inefrente:      string;
    ocr:            string;
    id_user:        number;
    operacion:      string;
    id_lider:       number;
}